+++ src/tablassert/log.py
@@ -1,18 +1,24 @@
 from __future__ import annotations
 
 from pathlib import Path
+from typing import TYPE_CHECKING
 
 from loguru import logger
+
+if TYPE_CHECKING:
+    from loguru import Logger
 
 LOGASSERT: Path = Path("./.logassert")
 LOGASSERT.mkdir(parents=True, exist_ok=True)
 
+LOG_FORMAT: str = "{time:YYYY-MM-DD HH:mm:ss} | {level} | {extra[category]} | {message}"
+
+logger.configure(extra={"category": "PIPELINE"})
 logger.remove()
 logger.add(
-    (LOGASSERT / "logassert.log"),
-    level="INFO",
-    format="{time:YYYY-MM-DD HH:mm:ss} | {level} | {message}",
-    rotation="100 MB",
-    encoding="utf-8",
-    mode="w",
+    (LOGASSERT / "logassert.log"), level="INFO", format=LOG_FORMAT, rotation="100 MB", encoding="utf-8", mode="w"
 )
+
+
+def cat(name: str) -> "Logger":
+    return logger.bind(category=name)